+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The official release will have a defined and more stable API. If you are already
  * Display status in output at the end of a build step (#4)
  * Display total build time in history (#16)
  * Indicate lost connection to LambdaCD in UI
+ * Allow retriggering of nested steps (#5)
+ * Redirect to new build when retriggering
 * API Changes:
  * removed deprecated argument lists for `lambdacd.internal.execution/execute-steps`